requirements.txt
influxdb==5.2.2
python-dateutil==2.8.0
RPi.GPIO==0.7.0
redis==3.3.6
